Change file handle submit to namespace submit handler #JC19-355
