+++ apps/vanilla/src/script/init.ts
@@ -1,11 +1,12 @@
-import { submitLoginForm, submitRegisterForm } from './handleSubmit';
+import { submitHandler } from '../namespaces/submitHandler';
+
 import { renderUserData } from './renderToUI';
 
 /** Init event for login form. */
 export function initLoginForm(): void {
   const form = document.querySelector('.form');
   if (form !== null) {
-    (form as HTMLFormElement).addEventListener('submit', submitLoginForm);
+    (form as HTMLFormElement).addEventListener('submit', submitHandler.submitLoginForm);
   }
 }
 
@@ -13,7 +14,7 @@ export function initLoginForm(): void {
 export function initRegisterForm(): void {
   const form = document.querySelector('.form');
   if (form !== null) {
-    (form as HTMLFormElement).addEventListener('submit', submitRegisterForm);
+    (form as HTMLFormElement).addEventListener('submit', submitHandler.submitRegisterForm);
   }
 }
